+++ src/level/BspRayTracer.hpp
@@ -15,7 +15,6 @@
 
 #pragma once
 
-#include "entity/component/BrushModelComponent.hpp"
 #include "math/FixedPoint.hpp"
 #include "level/BspLevel.hpp"
 #include "geo/Ray3.hpp"